Bugfix: Nonstatic alphabet order
There shouldn't be any duplicate characters in the alphabet file, so
converting to a set to eliminate duplicates isn't necessary.

By converting the list of characters to a set before making the alphabet
dictionary we lost the fixed ordering of the characters. The result was
that the characters would be encoded to different indices every time we
loaded the alphabet from the alphabet file.