Refactor data iteration and shuffling
Previously a BatchGenerator used a SampleGenerator to provide the
samples in a suitable order e.g. shuffled. Whenever the BatchGenerator
had enough samples for a batch it would process it and yield it.

Now a BatchGenerator will instead get a list of indices from an
IterationSchedule. The BatchGenerator itself will then be responsible
for fetching the samples with the given indices from the Loader.

The IterationSchedule is among other things given the desired batch size
and a reference to the Loader so it can do more "advanced" things like
splitting the dataset into batches without shuffling and then serving
those batches in shuffled order. Since the original dataset is sorted by
sequence lengths this results in serving batches where all sequences
have the same lengths, while still varying the sequence lengths between
batches.